Fixed a bug in the workaraound for Talk page retrieval
Sometimes, there are several writing variants (different upper/lower
case) for a Talk page, but only the commonly accepted writing variant
contains the actual content. The other writing variants redirect to the
main Talk page. When using a case insensitive DBMS setup, the API will
always return the set of all writing variants, but expects the result to
only consist of a single row. It will only use the first row returned.
If this is a redirect, we are stuck in an infinite loop. This patch does
not fix the issue, but prevent the loop from happening. It also adds
comments to the code with suggestions for a future fix and adds an
appropriate exception that is thrown in case of this happening.