Added methods to retrieve PK of first revision of an article
The rev PK is necessary for the RevisionIterator. It's possible to
retrieve the first revision by date and get the PK from there, but there
are cases when retrieval of first rev fails and returns some later
revision instead (e.g. article Third Reich in enwp). Retrieving the PK
of the first revision directly therefore allows simpler usage of the
RevisionIterator.

Also, added method that returns the RevisionAPIConfiguration object
similar to the Wikipedia class.