RevisionIterator was not loading all revision fields.
The RevisionIterator loads Revision objects differently from the
RevisionAPI. For some reason, it did not populate some of the fields
available for a revision data, such as contributor name, commit comment
etc. It is now changed to load the full revision objects. It is still
possible to avoid recovering the revision text if the corresponding
option is set.